+++ lib/design_system/all.rb
@@ -1,3 +1,10 @@
 require_relative 'generic'
 require_relative 'govuk'
 require_relative 'nhsuk'
+module DesignSystem
+  # Make zeitwerk happy by defining a stub class
+  # rubocop:disable Lint/EmptyClass
+  class All
+  end
+  # rubocop:enable Lint/EmptyClass
+end