+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Adding a stub class for all.rb to make zeitwerk happy
+
 ## [0.9.1] - 2025-10-08
 
 ### Fixed